Use [u]intptr_t instead of [unsigned] long to convert and/or represent pointers.
This finishes fixing conversions between pointers and integers of
possibly different sizes in GENERIC.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@38232 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/pci/scsiiom.c
@@ -1269,7 +1269,7 @@ NO_DEV:
 	    {
 		pACB->DeviceCnt++;
 		pPrevDCB = pDCB;
-		pACB->pDCB_free = (PDCB) ((ULONG) (pACB->pDCB_free) + sizeof( DC390_DCB ));
+		pACB->pDCB_free = (PDCB) ((uintptr_t) (pACB->pDCB_free) + sizeof( DC390_DCB ));
 		pDCB->DevType = bval1;
 		if(bval1 == SCSI_DASD || bval1 == SCSI_OPTICAL)
 		{
